[FIX] Retirar model de dentro do arquivo do contrato e remover alguns models
