+++ l10n_br_hr_payroll/models/__init__.py
@@ -15,8 +15,9 @@ from . import l10n_br_hr_rat_fap
 from . import l10n_br_hr_social_security_tax
 from . import resource_calendar
 from . import hr_salary_rule
-from . import l10n_br_hr_contract
+from . import hr_contract_change
 from . import hr_contract
+from . import hr_contribuicao_inss_vinculos
 from . import hr_contract_salary_rule
 from . import hr_department
 from . import hr_employee
